refactor(PersonNumber): 사람 수 리팩토링
